+++ plugins/DebuggerCore/unix/osx/PlatformEvent.cpp
@@ -20,17 +20,18 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "edb.h"
 #include <cstdio>
 #include <cstring>
+#include <signal.h> // for the SIG* definitions
 #include <sys/ptrace.h>
 #include <sys/types.h>
 #include <sys/wait.h>
-#include <signal.h> // for the SIG* definitions
 
 namespace DebuggerCore {
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-PlatformEvent::PlatformEvent() : status(0), pid(-1), tid(-1) {
+PlatformEvent::PlatformEvent()
+	: status(0), pid(-1), tid(-1) {
 }
 
 //------------------------------------------------------------------------------
@@ -49,38 +50,35 @@ IDebugEvent::Message PlatformEvent::error_description() const {
 	// TODO: figure out the fault address
 	const edb::address_t fault_address = 0;
 
-	switch(code()) {
+	switch (code()) {
 	case SIGSEGV:
 		return Message(
 			tr("Illegal Access Fault"),
 			tr(
 				"<p>The debugged application encountered a segmentation fault.<br />The address <strong>0x%1</strong> could not be accessed.</p>"
-				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>").arg(edb::v1::format_pointer(fault_address))
-			);
+				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>")
+				.arg(edb::v1::format_pointer(fault_address)));
 	case SIGILL:
 		return Message(
 			tr("Illegal Instruction Fault"),
 			tr(
 				"<p>The debugged application attempted to execute an illegal instruction.</p>"
-				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>")
-			);
+				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>"));
 	case SIGFPE:
 		// TODO: figure out the fault code for FPU stuff
-		switch(0) {
+		switch (0) {
 		case FPE_INTDIV:
-		return Message(
-			tr("Divide By Zero"),
-			tr(
-				"<p>The debugged application tried to divide an integer value by an integer divisor of zero.</p>"
-				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>")
-			);
+			return Message(
+				tr("Divide By Zero"),
+				tr(
+					"<p>The debugged application tried to divide an integer value by an integer divisor of zero.</p>"
+					"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>"));
 		default:
 			return Message(
 				tr("Floating Point Exception"),
 				tr(
 					"<p>The debugged application encountered a floating-point exception.</p>"
-					"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>")
-				);
+					"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>"));
 		}
 
 	case SIGABRT:
@@ -88,31 +86,27 @@ IDebugEvent::Message PlatformEvent::error_description() const {
 			tr("Application Aborted"),
 			tr(
 				"<p>The debugged application has aborted.</p>"
-				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>")
-			);
+				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>"));
 	case SIGBUS:
 		return Message(
 			tr("Bus Error"),
 			tr(
 				"<p>The debugged application tried to read or write data that is misaligned.</p>"
-				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>")
-			);
+				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>"));
 #ifdef SIGSTKFLT
 	case SIGSTKFLT:
 		return Message(
 			tr("Stack Fault"),
 			tr(
 				"<p>The debugged application encountered a stack fault.</p>"
-				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>")
-			);
+				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>"));
 #endif
 	case SIGPIPE:
 		return Message(
 			tr("Broken Pipe Fault"),
 			tr(
 				"<p>The debugged application encountered a broken pipe fault.</p>"
-				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>")
-			);
+				"<p>If you would like to pass this exception to the application press Shift+[F7/F8/F9]</p>"));
 	default:
 		return Message();
 	}
@@ -121,14 +115,14 @@ IDebugEvent::Message PlatformEvent::error_description() const {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-IDebugEvent::REASON PlatformEvent:: reason() const {
+IDebugEvent::REASON PlatformEvent::reason() const {
 	// this basically converts our value into a 'switchable' value for convenience
 
-	if(stopped()) {
+	if (stopped()) {
 		return EVENT_STOPPED;
-	} else if(terminated()) {
+	} else if (terminated()) {
 		return EVENT_TERMINATED;
-	} else if(exited()) {
+	} else if (exited()) {
 		return EVENT_EXITED;
 	} else {
 		return EVENT_UNKNOWN;
@@ -154,8 +148,8 @@ bool PlatformEvent::exited() const {
 // Name:
 //------------------------------------------------------------------------------
 bool PlatformEvent::is_error() const {
-	if(stopped()) {
-		switch(code()) {
+	if (stopped()) {
+		switch (code()) {
 		case SIGTRAP:
 		case SIGSTOP:
 			return false;
@@ -229,16 +223,16 @@ edb::tid_t PlatformEvent::thread() const {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-int PlatformEvent::code() const {
-	if(stopped()) {
+int64_t PlatformEvent::code() const {
+	if (stopped()) {
 		return WSTOPSIG(status);
 	}
 
-	if(terminated()) {
+	if (terminated()) {
 		return WTERMSIG(status);
 	}
 
-	if(exited()) {
+	if (exited()) {
 		return WEXITSTATUS(status);
 	}
